Renamed backgrounds to background
