+++ lib/background/background.ts
